PAE-166: refactor endpoints to remove code duplication
