+++ src/routes/v1/apply/registration.js
@@ -1,19 +1,8 @@
-import Boom from '@hapi/boom'
-import {
-  extractAnswers,
-  extractOrgId,
-  extractReferenceNumber
-} from '../../../common/helpers/apply/extract-answers.js'
+import { registrationAndAccreditationHandler } from '../../../common/helpers/apply/handler.js'
+import { registrationAndAccreditationPayload } from '../../../common/helpers/apply/payload.js'
 import { registrationFactory } from '../../../common/helpers/collections/factories/index.js'
-import { createLogger } from '../../../common/helpers/logging/logger.js'
-import {
-  LOGGING_EVENT_ACTIONS,
-  LOGGING_EVENT_CATEGORIES
-} from '../../../common/enums/event.js'
 
-const path = '/v1/apply/registration'
-
-const registrationPath = path
+const registrationPath = '/v1/apply/registration'
 
 /**
  * Apply: Registration
@@ -21,68 +10,17 @@ const registrationPath = path
  */
 const registration = {
   method: 'POST',
-  path,
+  path: registrationPath,
   options: {
     validate: {
-      payload: (data, _options) => {
-        if (!data || typeof data !== 'object') {
-          throw Boom.badRequest('Invalid payload')
-        }
-
-        const answers = extractAnswers(data)
-        const orgId = extractOrgId(answers)
-        const referenceNumber = extractReferenceNumber(answers)
-
-        if (!orgId) {
-          throw Boom.badRequest('Could not extract orgId from answers')
-        }
-
-        if (!referenceNumber) {
-          throw Boom.badRequest(
-            'Could not extract referenceNumber from answers'
-          )
-        }
-
-        return { answers, orgId, rawSubmissionData: data, referenceNumber }
-      }
+      payload: registrationAndAccreditationPayload
     }
   },
-  handler: async ({ db, payload }, h) => {
-    const { answers, orgId, rawSubmissionData, referenceNumber } = payload
-    const logger = createLogger()
-
-    try {
-      await db.collection('registration').insertOne(
-        registrationFactory({
-          orgId,
-          referenceNumber,
-          answers,
-          rawSubmissionData
-        })
-      )
-
-      logger.info({
-        message: `Stored registration data for orgId: ${orgId} and referenceNumber: ${referenceNumber}`,
-        event: {
-          category: LOGGING_EVENT_CATEGORIES.SERVER,
-          action: LOGGING_EVENT_ACTIONS.REQUEST_SUCCESS
-        }
-      })
-      return h.response().code(201)
-    } catch (err) {
-      const message = `Failure on ${registrationPath}`
-
-      logger.error(err, {
-        message,
-        event: {
-          category: LOGGING_EVENT_CATEGORIES.SERVER,
-          action: LOGGING_EVENT_ACTIONS.RESPONSE_FAILURE
-        }
-      })
-
-      throw Boom.badImplementation(message)
-    }
-  }
+  handler: registrationAndAccreditationHandler(
+    'accreditation',
+    registrationPath,
+    registrationFactory
+  )
 }
 
 export { registration, registrationPath }
